Sync

+++ src/FastGraph.h
@@ -6,15 +6,15 @@
 #include "utils.h"
 
 
-typedef unsigned int vertex;
+typedef unsigned long long vertex;
 
 
 class FastGraph {
  private:
     bool weighted;
     std::vector< vertex > vertices;
-    std::unordered_map< vertex, int > renumerationTable;//Converts from new indices to old
-    std::vector< int > indices;
+    std::unordered_map< vertex, unsigned long long > renumerationTable;//Converts from new indices to old
+    std::vector< unsigned long long > indices;
     std::vector< vertex > csr;
     std::vector< double > weights;
     std::vector< double > scores;
@@ -22,7 +22,7 @@ class FastGraph {
 
     std::vector< std::vector< double > > threadScores;
 
-    void threadFunction(unsigned int id, unsigned int finishIndex, std::atomic<unsigned int>& counter, std::atomic<unsigned int>& finishedTasks, std::atomic<bool>& shouldBeRunning, std::atomic<unsigned int>& runningThreads);
+    void threadFunction(unsigned int id, unsigned long long int finishIndex, std::atomic<unsigned long long>& counter, std::atomic<unsigned long long>& finishedTasks, std::atomic<bool>& shouldBeRunning, std::atomic<unsigned int>& runningThreads);
 
  public:
     FastGraph(globalArgs_t *_args):
